- Remove MD files covered by API3 Docs - Move types to base directory - Export types from index - Update index and types reference in package.json to enable cleaner imports
