+++ package.json
@@ -4,8 +4,8 @@
   "private": false,
   "version": "0.0.1-d",
   "description": "A package to parse API3 operations metadata",
-  "main": "./dist/index",
-  "types": "./dist/validation/types",
+  "main": "./dist/index.js",
+  "types": "./dist/index.js",
   "files": [
     "dist",
     "data",
