+++ src/deployment/deploy-airkeeper.ts
@@ -1,6 +1,6 @@
 import { join } from 'path';
 import { PromptObject } from 'prompts';
-import { OperationsRepository } from '../validation/types';
+import { OperationsRepository } from '../types';
 import { cliPrint, runAndHandleErrors, runShellCommand } from '../utils/cli';
 import { promptQuestions } from '../utils/prompts';
 import { readAndValidateOperationsRepository } from '../utils/read-operations';
