+++ src/deployment/test-deployment.ts
@@ -2,7 +2,7 @@ import { existsSync } from 'fs';
 import { join } from 'path';
 import axios, { AxiosRequestConfig } from 'axios';
 import { PromptObject } from 'prompts';
-import { OperationsRepository } from '../validation/types';
+import { OperationsRepository } from '../types';
 import { cliPrint, runAndHandleErrors } from '../utils/cli';
 import { promptQuestions } from '../utils/prompts';
 import { readOperationsRepository } from '../utils/read-operations';
