+++ src/register/enable-policy-readers.ts
@@ -1,6 +1,6 @@
 import { NonceManager } from '@ethersproject/experimental';
 import { ContractTransaction, ethers } from 'ethers';
-import { Policy } from '../validation/types';
+import { Policy } from '../types';
 import { runAndHandleErrors } from '../utils/cli';
 import { getDapiNameHash, getDapiServerContract } from '../utils/evm';
 import { loadCredentials } from '../utils/filesystem';
