+++ src/utils/generate-topup-wallets.ts
@@ -6,7 +6,7 @@ import { readOperationsRepository } from './read-operations';
 import { writeOperationsRepository } from './write-operations';
 import { runAndHandleErrors } from './cli';
 import { PROTOCOL_ID_PSP } from './evm';
-import { OperationsRepository } from '../validation/types';
+import { OperationsRepository } from '../types';
 
 const questions = (choices: Choice[]): PromptObject[] => {
   return [
