+++ test/deployment/create-airseeker-config.test.ts
@@ -5,7 +5,7 @@ import * as prompts from 'prompts';
 import { readOperationsRepository } from '../../src/utils/read-operations';
 import { createAirseekerConfig } from '../../src/deployment/create-airseeker-config';
 import { writeOperationsRepository } from '../../src/utils/write-operations';
-import { OperationsRepository } from '../../src/validation/types';
+import { OperationsRepository } from '../../src/types';
 import { getFormattedUtcTimestamp } from '../../src/utils/date';
 
 describe('create-airseeker-config', () => {
