+++ test/deployment/create-config.test.ts
@@ -6,7 +6,7 @@ import * as createConfigModule from '../../src/deployment/create-config';
 import * as dateModule from '../../src/utils/date';
 import { readOperationsRepository } from '../../src/utils/read-operations';
 import { writeOperationsRepository } from '../../src/utils/write-operations';
-import { OperationsRepository } from '../../src/validation/types';
+import { OperationsRepository } from '../../src/types';
 import { getFormattedUtcTimestamp } from '../../src/utils/date';
 
 describe('create-config', () => {
